Add CI for CTest and clang-tidy (#11)
* Add CI runs for Clang Tidy

* Use ARM for GitHub runners

* Disable building of sorting homework due to assembly architecture issues

* Run tests using CTest

* Add .clang-tidy

* Update clang-tidy

* Fix basic clang-tidy warnings

* Attempt to fix uninitialized value warning

* Include assert

* Use calloc for allocation

* Use operators instead of operator

* Fix clang-tidy warnings

* Format

* Fix char rounding

* Fix job name

+++ hws/substring/main.c
@@ -9,17 +9,17 @@ int main(void)
     char* s = promptString("S: ");
     char* s1 = promptString("S1: ");
 
-    int lenS = strlen(s);
-    int lenS1 = strlen(s1);
+    size_t lenS = strlen(s);
+    size_t lenS1 = strlen(s1);
 
     int n = 0;
 
     if (lenS == lenS1 && strcmp(s, s1) == 0) {
         n = 1;
     } else if (lenS > lenS1) {
-        for (int i = 0; i <= lenS - lenS1; i++) {
+        for (size_t i = 0; i <= lenS - lenS1; i++) {
             bool ok = true;
-            for (int j = 0; j < lenS1; j++) {
+            for (size_t j = 0; j < lenS1; j++) {
                 if (s1[j] != s[j + i]) {
                     ok = false;
                     break;
